custom_image/flask_server/requirements.txt
Flask
flasgger
pytz
python-dotenv
flask-sqlalchemy
psycopg2-binary
flask-swagger
pandas